Move relationships code out
Into its own class
