+++ lib/json_api/resource.rb
@@ -1,9 +1,11 @@
 require_relative 'resources/base'
+require_relative 'resources/relationships'
 require_relative 'resources/dsl'
 
 module JSONAPI
   class Resource
     include Resources::Base
+    include Resources::Relationships
     extend Resources::DSL
 
     def self.inherited(subclass)
